+++ tests/testthat/test-python-pandas.R
@@ -15,11 +15,51 @@ test_that("Simple Pandas data frames can be roundtripped", {
 
 test_that("Ordered factors are preserved", {
   skip_if_no_pandas()
+  
   pd <- import("pandas")
   
   set.seed(123)
   before <- data.frame(x = ordered(letters, levels = sample(letters)))
   after <- py_to_r(r_to_py(before))
   expect_equal(before, after)
+  
+})
+
+test_that("Timestamped arrays in Pandas DataFrames can be roundtripped", {
+  skip_if_no_pandas()
+  
+  pd <- import("pandas", convert = FALSE)
+  np <- import("numpy", convert = FALSE)
+  
+  data <- list(
+    'A' = 1.,
+    'B' = pd$Timestamp('20130102'),
+    'C' = pd$Series(1:4, dtype = 'float32'),
+    'D' = np$array(rep(3L, 4), dtype = 'int32'),
+    'E' = pd$Categorical(c("test", "train", "test", "train")),
+    'F' = 'foo'
+  )
+  
+  before <- pd$DataFrame(data)
+  
+  converted <- py_to_r(before)
+  
+  after <- r_to_py(converted)
+  
+  expect_equal(py_to_r(before$to_csv()), py_to_r(after$to_csv()))
+  
+})
+
+test_that("data.frames with length-one factor columns can be converted", {
+  skip_if_no_pandas()
+  
+  pd <- import("pandas", convert = FALSE)
+  np <- import("numpy", convert = FALSE)
+  
+  before <- data.frame(x = "hello")
+  converted <- r_to_py(before)
+  after <- py_to_r(converted)
+  
+  expect_identical(before, after)
   
 })